@@ -2,8 +2,8 @@
 using System.Linq;
 using System.Threading.Tasks;
 using ArchitectNow.ApiStarter.Common.Models.Options;
+using ArchitectNow.ApiStarter.Common.Models.Security;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
-using Microsoft.AspNetCore.Builder;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.IdentityModel.Tokens;
@@ -12,10 +12,14 @@ namespace ArchitectNow.ApiStarter.Api.Configuration
 {
     public static class JwtExtensions
     {
-        public static void ConfigureJwt(this IServiceCollection services, IConfiguration configuration,
-            Func<JwtIssuerOptions, SecurityKey> signingKey)
+        public static void ConfigureJwt(this IServiceCollection services, IConfiguration configurationRoot,
+            Func<JwtIssuerOptions, JwtSigningKey> signingKey, JwtBearerEvents jwtBearerEvents = null)
         {
-            var jwtAppSettingOptions = configuration.GetSection(nameof(JwtIssuerOptions)).Get<JwtIssuerOptions>();
+            var jwtAppSettingOptions = configurationRoot.GetSection(nameof(JwtIssuerOptions)).Get<JwtIssuerOptions>();
+
+            var issuerSigningKey = signingKey(jwtAppSettingOptions);
+
+            services.AddSingleton(issuerSigningKey);
 
             var tokenValidationParameters = new TokenValidationParameters
             {
@@ -26,7 +30,7 @@ namespace ArchitectNow.ApiStarter.Api.Configuration
                 ValidAudience = jwtAppSettingOptions.Audience,
 
                 ValidateIssuerSigningKey = true,
-                IssuerSigningKey = signingKey(jwtAppSettingOptions),
+                IssuerSigningKey = issuerSigningKey,
 
                 RequireExpirationTime = true,
                 ValidateLifetime = true,
@@ -57,11 +61,6 @@ namespace ArchitectNow.ApiStarter.Api.Configuration
                         }
                     };
                 });
-        }
-
-        public static void ConfigureJwt(this IApplicationBuilder app)
-        {
-            app.UseAuthentication();
         }
     }
 }
