@@ -1,10 +1,9 @@
-﻿using ArchitectNow.ApiStarter.Api.Filters;
+﻿using System;
+using ArchitectNow.ApiStarter.Api.Filters;
+using ArchitectNow.ApiStarter.Api.Models.Validation;
 using FluentValidation.AspNetCore;
-using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Mvc.Authorization;
+using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.DependencyInjection;
-using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
 using Newtonsoft.Json.Serialization;
 
@@ -12,28 +11,39 @@ namespace ArchitectNow.ApiStarter.Api.Configuration
 {
     public static class WebApiExtensions
     {
-        public static void ConfigureApi(this IServiceCollection services)
+        public static void ConfigureApi(this IServiceCollection services,
+            FluentValidationOptions fluentValidationOptions, Action<MvcOptions> configureMvc = null,
+            Action<MvcJsonOptions> configureJson = null)
         {
+            /*************************
+             * IConfiguration is not available yet
+             *************************/
+
             services.AddAntiforgery(options => options.HeaderName = "X-XSRF-TOKEN");
-            services.AddRouting(options => options.LowercaseUrls = true);
 
-            services.AddCors((options => options.AddPolicy("AllowAllOrigins",
-                builder =>
-                {
-                    builder.AllowAnyOrigin();
-                })));
+            services.AddRouting(options =>
+            {
+                options.LowercaseUrls = true;
+                options.LowercaseQueryStrings = true;
+            });
 
-            services.AddMvc(o =>
+            services.AddApiVersioning(options =>
+            {
+                options.AssumeDefaultVersionWhenUnspecified = true;
+                options.DefaultApiVersion = new ApiVersion(1, 0);
+                options.ReportApiVersions = true;
+                options.UseApiBehavior = false;
+            });
+
+            var mvcBuilder = services.AddMvcCore(o =>
                 {
                     o.Filters.AddService(typeof(GlobalExceptionFilter));
                     o.ModelValidatorProviders.Clear();
 
-                    var policy = new AuthorizationPolicyBuilder()
-                        .RequireAuthenticatedUser()
-                        .Build();
-
-                    o.Filters.Add(new AuthorizeFilter(policy));
+                    configureMvc?.Invoke(o);
                 })
+                .AddAuthorization()
+                .AddJsonFormatters()
                 .AddJsonOptions(options =>
                 {
                     var settings = options.SerializerSettings;
@@ -41,22 +51,21 @@ namespace ArchitectNow.ApiStarter.Api.Configuration
                     var camelCasePropertyNamesContractResolver = new CamelCasePropertyNamesContractResolver();
 
                     settings.ContractResolver = camelCasePropertyNamesContractResolver;
-                    settings.Converters = new JsonConverter[]
-                    {
-                        new IsoDateTimeConverter(),
-                        new StringEnumConverter(true)
-                    };
-                })
-                .AddFluentValidation();
-        }
+                    settings.Converters.Add(new IsoDateTimeConverter());
+                    settings.Converters.Add(new StringEnumConverter(new DefaultNamingStrategy()));
 
-        public static void ConfigureAssets(this IApplicationBuilder app)
-        {
-            app.UseFileServer();
+                    configureJson?.Invoke(options);
+                });
 
-            app.UseStaticFiles();
+            services.AddVersionedApiExplorer(options =>
+            {
+                options.GroupNameFormat = "VVV";
+                options.SubstituteApiVersionInUrl = true;
+            });
 
-            app.UseCors("AllowAllOrigins");
+            if (fluentValidationOptions.Enabled)
+                mvcBuilder.AddFluentValidation(
+                    configuration => fluentValidationOptions.Configure?.Invoke(configuration));
         }
     }
 }
