@@ -9,23 +9,19 @@ namespace ArchitectNow.ApiStarter.Api.Services
     {
         public static UserInformation GetUserInformation(this ClaimsPrincipal principal)
         {
-            if (principal == null)
-                return null;
+            if (principal == null) return null;
 
-            var identityName = principal?.Identity?.Name;
-            var role = principal?.Claims.FirstOrDefault(claim =>
-                string.Equals(nameof(UserInformation.UserRole), claim.Type, StringComparison.OrdinalIgnoreCase))?.Value;
-            var idValue = principal?.Claims.FirstOrDefault(claim =>
+            var identityName = principal.Identity?.Name;
+
+            var idValue = principal.Claims.FirstOrDefault(claim =>
                 string.Equals(nameof(UserInformation.Id), claim.Type, StringComparison.OrdinalIgnoreCase))?.Value;
 
             var userInformation = new UserInformation
             {
-                Email = identityName,
-                UserRole = role
+                Email = identityName
             };
 
-            if (Guid.TryParse(idValue, out var id))
-                userInformation.Id = id;
+            if (Guid.TryParse(idValue, out var id)) userInformation.Id = id;
 
             return userInformation;
         }
