@@ -11,8 +11,7 @@ namespace ArchitectNow.ApiStarter.Common.Models.Validation
             RuleFor(vm => vm.Email).NotEmpty().EmailAddress();
             RuleFor(vm => vm.NameFirst).MinimumLength(3);
             RuleFor(vm => vm.NameLast).MinimumLength(3);
-            RuleFor(vm => vm.Password).Password().WithMessage("Password must be 4 numbers only");
-
+            RuleFor(vm => vm.Password).Password().WithMessage("Password must be 6 numbers only");
         }
     }
 }
