@@ -15,10 +15,10 @@ namespace ArchitectNow.ApiStarter.Common.Repositories
     public class PersonRepository : BaseRepository<Person>, IPersonRepository
     {
         public PersonRepository(ILogger<Person> logger,
-            IDataContext dataContext, ICacheService cacheService,
+            ICacheService cacheService,
             IOptions<MongoOptions> options,
             IValidator<Person> validator = null
-        ) : base(logger, dataContext, cacheService, options, validator)
+        ) : base(logger, cacheService, options, validator)
         {
         }
 
@@ -31,8 +31,9 @@ namespace ArchitectNow.ApiStarter.Common.Repositories
             var collection = GetCollection();
 
             await collection.Indexes.CreateOneAsync(
-                Builders<Person>.IndexKeys.Ascending(l => l.NameLast),
-                new CreateIndexOptions {Name = "person_nameLast"});
+                new CreateIndexModel<Person>(
+                    Builders<Person>.IndexKeys.Ascending(l => l.NameLast),
+                    new CreateIndexOptions {Name = "person_nameLast"}));
         }
 
         public async Task<List<Person>> Search(string searchString)
